Updated namespaces in moved files. Removed references on non-existing namespaces.
