+++ NetworkConnections/Gateway/Entities/Channels/DM/GroupDMTextChannel.cs
@@ -6,7 +6,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Gateway.Entities.Channels.Text
+namespace Gateway.Entities.Channels.DM
 {
     [JsonObject(MemberSerialization.OptIn)]
     internal class GroupDMTextChannel : DMTextChannel
