+++ NetworkConnections/Gateway/Entities/Channels/Guild/GuildStoreChannel.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Gateway.Entities.Channels.Text
+namespace Gateway.Entities.Channels.Guild
 {
     [JsonObject(MemberSerialization.OptIn)]
     internal class GuildStoreChannel : GuildTextChannelBase
