+++ NetworkConnections/Gateway/Entities/Channels/Guild/IGuildChannel.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace Gateway.Entities.Channels
+namespace Gateway.Entities.Channels.Guild
 {
     public interface IGuildChannel : IChannel
     {
