+++ NetworkConnections/Gateway/Entities/Channels/Guild/IUpdatable/IUpdatableGuildCategory.cs
@@ -1,4 +1,4 @@
-﻿namespace Gateway.Entities.Channels
+﻿namespace Gateway.Entities.Channels.Guild.IUpdatable
 {
     public interface IUpdatableGuildCategory : IUpdatableGuildChannel
     {
