+++ NetworkConnections/Gateway/Entities/Channels/Guild/Text/GuildTextChannel.cs
@@ -6,7 +6,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Gateway.Entities.Channels.Text
+namespace Gateway.Entities.Channels.Guild.Text
 {
     [JsonObject(MemberSerialization.OptIn)]
     internal class GuildTextChannel : GuildTextChannelBase
