+++ NetworkConnections/Gateway/Entities/Channels/Guild/Text/GuildTextChannelBase.cs
@@ -6,7 +6,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Gateway.Entities.Channels.Text
+namespace Gateway.Entities.Channels.Guild.Text
 {
     [JsonObject(MemberSerialization.OptIn)]
     internal abstract class GuildTextChannelBase : TextChannel, IGuildChannel//, IUpdatableGuildTextChannel
